+++ ha-addon/calendhd/rootfs/opt/calendhd/public/_app/immutable/entry/start.RvHoYNcc.js
@@ -1,0 +1,1 @@
+import{c as e,r as t}from"../chunks/2HyLElpo.js";export{e as load_css,t as start};